+++ src/7segment/_7segment_leds.h
@@ -1,5 +1,5 @@
 #include <Arduino.h>
-#include "definitions.h"
+#include "settings.h"
 #if SEVEN_SEGMENT
 #define SEGMENT_LEDS 8
 const int leds[SEGMENT_LEDS] PROGMEM = {2, 3, 4, 5, 6, 7, 8, 9};