+++ src/UKP/UKPmanager.cpp
@@ -4,14 +4,6 @@
 #include "macrosengine/macrosengine.hpp"
 #endif
 
-#if !LED
-#ifdef LED_BUILTIN
-#define LED_PIN LED_BUILTIN
-#endif
-#else 
-#define LED_PIN LED
-#endif
-
 UKPmanager manager;
 /**
  * @brief Initialize all the components needed to di