+++ src/button/button_pins.h
@@ -1,5 +1,5 @@
 #include <Arduino.h>
-#include "definitions.h"
+#include "settings.h"
 /*Define the pins for each of the number of BUTTONS declared in definitions.h
 (Note that the first number button pin is used for profile switching 
 (unless PROFILE is set to 0)) */